move Ownable tests into AccessManager.test.js
